feat: Add handler layer for Project and Task with repository and logic updates
-Added handler layer for managing Project and Task entities.
-Implemented a method in the Project service to retrieve all projects by
a specified userID.
-Added a method in the TaskRepository to fetch all tasks by a specified
projectID.
-Made minor adjustments to the logic of certain existing methods for
improved functionality and consistency.

+++ internal/services/project_service.go
@@ -18,6 +18,9 @@ func (p *ProjectService) CreateProject(project *models.Project) error {
 	if project.Name == "" {
 		return errors.New("project name can not be empty")
 	}
+	if project.UserID == 0 {
+		return errors.New("project must associated with a user")
+	}
 	return p.ProjectRepo.CreateProject(project)
 }
 
@@ -26,9 +29,6 @@ func (p *ProjectService) GetProjectById(id uint) (*models.Project, error) {
 	if err != nil {
 		return nil, err
 	}
-	if project == nil {
-		return nil, errors.New("project not found")
-	}
 	return project, nil
 }
 
@@ -36,13 +36,31 @@ func (p *ProjectService) UpdateProject(project *models.Project) error {
 	if project.Name == "" {
 		return errors.New("project name can not be empty")
 	}
+	projectUpdated, err := p.ProjectRepo.GetProjectById(project.ID)
+	if err != nil {
+		return err
+	}
+	if projectUpdated == nil {
+		return errors.New("project doesn't exist")
+	}
 	return p.ProjectRepo.UpdateProject(project)
 }
 
 func (p *ProjectService) DeleteProjectById(id uint) error {
+	project, err := p.ProjectRepo.GetProjectById(id)
+	if err != nil {
+		return err
+	}
+	if project == nil {
+		return errors.New("project doesn't exist")
+	}
 	return p.ProjectRepo.DeleteProjectById(id)
 }
 
-func (p *ProjectService) GetAllProjects() ([]models.Project, error) {
+func (p *ProjectService) GetAllProjects() ([]*models.Project, error) {
 	return p.ProjectRepo.GetAllProjects()
+}
+
+func (p *ProjectService) GetAllProjectByUserId(userId uint) ([]*models.Project, error) {
+	return p.ProjectRepo.GetAllProjectByUserId(userId)
 }
